requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
sentence-transformers
pinecone
langchain-core
langchain-community
langchain-text-splitters
groq
langchain

python-multipart